pyproject.toml
[tool.poetry]
name = "repl_python3_Raider-Bots-Py"
version = "0.1.0"
description = ""
authors = ["Mitch Schwartz - TW: @MitchSchwartzCo"]

[tool.poetry.dependencies]
python = "^3.8"
Flask = "^2.0.2"
"discord.py" = "^1.7.3"
dateparser = "^1.1.0"
pytz = "^2021.3"
django-timezone = "^0.3"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
